refactor: 유튜브 관련 비즈니스 로직을 hooks로 리팩토링.
